requirements.txt
aiogrpc >= 1.6
cryptography >= 2.7
hkdf >= 0.0.3
pycryptodomex >= 3.8.2
pysha3 >= 1.0.2
requests >= 2.22.0
six >= 1.12.0
protobuf >= 3.9.0
pyyaml >= 5.1
dacite >= 1.0.2